+++ src/main/java/shop/tryit/domain/order/OrderDetailRepository.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Optional;
 import shop.tryit.domain.order.dto.OrderDetailSearchDto;
 import shop.tryit.domain.order.entity.Order;
+import shop.tryit.domain.order.entity.OrderDetail;
 
 public interface OrderDetailRepository {
