+++ src/main/java/shop/tryit/domain/order/service/OrderFacade.java
@@ -10,11 +10,11 @@ import shop.tryit.domain.item.entity.Item;
 import shop.tryit.domain.item.service.ItemService;
 import shop.tryit.domain.member.Member;
 import shop.tryit.domain.member.MemberService;
-import shop.tryit.domain.order.OrderDetail;
 import shop.tryit.domain.order.OrderStatus;
 import shop.tryit.domain.order.dto.OrderDetailDto;
 import shop.tryit.domain.order.dto.OrderSearchDto;
 import shop.tryit.domain.order.entity.Order;
+import shop.tryit.domain.order.entity.OrderDetail;
 
 @Component
 @RequiredArgsConstructor
